added MessageBox "Are you sure you want to delete the selected item(s)?" in case user wants to remove item (single or group)

+++ ShortCuts Manager/ShortCuts Manager/MainWindowViewModel.cs
@@ -212,10 +212,18 @@ namespace ShortCuts_Manager
 
         public void DeleteSingle()
         {
-            foreach (var info in SelectedSingleShortCutInformation.ToList())
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to delete the selected item(s)?",
+                "Confirm Deletion",
+                MessageBoxButton.YesNo,
+                MessageBoxImage.Warning);
+
+            if (result == MessageBoxResult.Yes)
             {
-                SingleShortCutInformation.Remove(info);
-                dataBase.RemoveSingle(info);
+                foreach (var info in SelectedSingleShortCutInformation.ToList())
+                {
+                    SingleShortCutInformation.Remove(info);
+                    dataBase.RemoveSingle(info);
+                }
             }
         }
         #endregion DeleteSingle
@@ -226,8 +234,16 @@ namespace ShortCuts_Manager
 
         public void DeleteGroup()
         {
-            dataBase.RemoveGroup(SelectedGroupShortCutsInformation);
-            GroupShortCutsInformation.Remove(SelectedGroupShortCutsInformation);
+            MessageBoxResult result = MessageBox.Show("Are you sure you want to delete the selected item(s)?",
+                "Confirm Deletion",
+                MessageBoxButton.YesNo,
+                MessageBoxImage.Warning);
+
+            if (result == MessageBoxResult.Yes)
+            {
+                dataBase.RemoveGroup(SelectedGroupShortCutsInformation);
+                GroupShortCutsInformation.Remove(SelectedGroupShortCutsInformation);
+            }
         }
         #endregion DeleteGroup
